requirements.txt
openai==1.30.1
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn==0.29.0
httpx==0.28.1
pydantic==2.11.1
python-multipart
faster-whisper @ git+https://github.com/SYSTRAN/faster-whisper.git
gTTS==2.5.1
streamlit
playsound==1.2.2
google-cloud-texttospeech

